requirements.txt
Flask==2.2.5
flask-cors==3.0.10
openai==0.27.0
faiss-cpu==1.7.2
sentence-transformers==2.2.2
PyPDF2==3.0.1
gunicorn==20.1.0
